update trade schema
The utilization of multiple fields for storing the price of a specific trade type is deemed unnecessary. Therefore, I have removed the "maybe_price" and "unit_price" fields, and we now employ the "price" field for all trade types. Additionally, I have included the "highest_bid" field to store the highest bid in auction trades.